requirements.txt
langchain~=0.3.3
langchain_openai~=0.2.2
python-dotenv~=1.0.1
requests~=2.32.3
python-dateutil~=2.9.0.post0
tqdm~=4.66.5
langchain-core~=0.3.10
langchain_openai~=0.2.2
pandas~=2.2.3
pycountry~=24.6.1
pycountry-convert~=0.7.2
aiohttp~=3.10.10
aiofiles~=24.1.0
rich~=13.9.2
pyyaml~=6.0.2
black~=25.1.0